requirements.txt
jinja2==3.1.2
python-dotenv==1.0.0
pdfkit==1.0.0
pandas==2.2.3
openpyxl==3.1.2
flask==3.0.2